refactor(flake): extract shared vitest-list resolver for the two guards (#916)
/simplify pass on the sprint diff. Two cleanups:

- check-no-page-mounts.mjs: drop a redundant double `/(^|\/)App$/` test in
  isPageMountSpecifier — the basename pop ('App') already covers every relative
  form (../App, ./App, .../App), so the early-return regex was dead.
- Extract the duplicated `vitest list --json` exec+parse (was copied between
  check-test-projects.mjs and check-no-page-mounts.mjs) into
  frontend/scripts/lib/listVitestFiles.mjs. Both guards now source the unit/
  integration file set from one resolver, so they can't drift in HOW they ask
  vitest (R20/R22 single-source spirit). Net -67 lines.

Pure mechanical extraction: the set logic (disjoint/exhaustive, page-mount
scan) is unchanged from the green pre-refactor run (239 = 176 unit + 63
integration; no-page-mounts OK). Modules load + syntax-check clean.
